// cypress/cypress/support/helpers/enums/navbar-enums.ts
export enum NavBarVars {
  navBar = 'navbar',
  logoBtn = 'logobtn',
  liveSlotsBtn = 'navbar-live-slots',
  liveCasinoBtn = 'navbar-live-casino',
  eGamesBtn = 'navbar-e-games',
  offers = 'navbar-offers',
  rewardsBtn = 'navbar-rewards',
  regBtn = 'regbtn',
  loginBtn = 'loginbtn',
  langBtn = 'langbtn',
  langDropdown = 'lang-dropdown',
  engOption = 'English',
  chinaOption = '繁體中文',
  koreanOption = '한국어'
}