// cypress/cypress/support/helpers/enums/url-enum.ts
export enum urlText {
  register = 'register',
  login = 'login',
  liveCasino = 'live-casino',
  liveSlots = 'live-slots',
  classicLiveSlots = 'slots-classic',
  eGames = 'egames',
  rewards = 'rewards',
  solaireResort = 'solaireresort',
  baccaratCasino = 'baccarat',
  stadium = 'stadium',
  jinJi = 'JIN_JI_BAO_XI',
  depo = 'deposit',
  megaways = 'megaways',
  terms = 'terms-conditions',
  jinJi50m = 'jin_ji_bao_xi_link_50m',
  jackpots = 'jackpots',
  en = 'en',
  kr = 'kr',
  cn = 'cn'
}

export enum urlLiveCasino {
  baccarat = 'baccarat',
  stadium = 'stadium',
  sicbo = 'sicbo',
  roulette = 'roulette'
}

export enum urleGames {
  new = 'new',
  // ancientCivilizations = 'ancient-civilizations',
  table = 'table',
  evoplay = 'evoplay'
  // megaways = 'megaways',
  // classicStyle = 'classic-style',
  // allGames = 'all'
}

export enum urlProfile {
  profile = 'profile',
  balance = 'balance',
  transactions = 'transactions',
  settings = 'settings',
  responsibleGaming = 'responsible-gaming'
}